Fixes for invoice page banner and line items
