+++ src/frontend/extract-helpers/get-invoice-banner.ts
@@ -1,8 +1,8 @@
 import {
-  DiscrepancySpecialCase,
+  PercentageDiscrepancySpecialCase,
   InvoiceBannerClass,
   InvoiceBannerStatus,
-  percentageDiscrepancySpecialCase,
+  percentageDiscrepancySpecialCases,
 } from "../utils";
 import { FullExtractLineItem } from "./types";
 
@@ -14,10 +14,10 @@ interface InvoiceBanner {
 // Note that these are just the special cases that we want to show a warning for --
 // MN_NO_CHARGE is also a possible value in a line item but, it doesn't cause a warning.
 const WARNINGS_BY_PRIORITY = [
-  percentageDiscrepancySpecialCase.MN_INVOICE_MISSING,
-  percentageDiscrepancySpecialCase.MN_EVENTS_MISSING,
-  percentageDiscrepancySpecialCase.MN_RATES_MISSING,
-  percentageDiscrepancySpecialCase.MN_UNEXPECTED_CHARGE,
+  percentageDiscrepancySpecialCases.MN_INVOICE_MISSING,
+  percentageDiscrepancySpecialCases.MN_EVENTS_MISSING,
+  percentageDiscrepancySpecialCases.MN_RATES_MISSING,
+  percentageDiscrepancySpecialCases.MN_UNEXPECTED_CHARGE,
 ];
 
 const WARNING_CODES = WARNINGS_BY_PRIORITY.map(
@@ -32,7 +32,7 @@ const anyHaveWarning = (lineItems: FullExtractLineItem[]): boolean => {
 
 const findHighestPriorityWarning = (
   lineItems: FullExtractLineItem[]
-): DiscrepancySpecialCase => {
+): PercentageDiscrepancySpecialCase => {
   const highestPriorityWarning = WARNINGS_BY_PRIORITY.find((warning) =>
     lineItems.find(
       (lineItem) => lineItem.price_difference_percentage === warning.magicNumber
@@ -56,7 +56,7 @@ const anyAreBelowThreshold = (lineItems: FullExtractLineItem[]): boolean => {
     ({ price_difference_percentage }) =>
       +price_difference_percentage <= -1 &&
       price_difference_percentage !==
-        percentageDiscrepancySpecialCase.MN_NO_CHARGE.magicNumber
+        percentageDiscrepancySpecialCases.MN_NO_CHARGE.magicNumber
   );
 };
 
@@ -64,7 +64,7 @@ const anyAreNotNoCharge = (lineItems: FullExtractLineItem[]): boolean => {
   return lineItems.some(
     (lineItem) =>
       lineItem.price_difference_percentage !==
-      percentageDiscrepancySpecialCase.MN_NO_CHARGE.magicNumber
+      percentageDiscrepancySpecialCases.MN_NO_CHARGE.magicNumber
   );
 };
 
@@ -85,7 +85,7 @@ export const getInvoiceBanner = (
     bannerClass = InvoiceBannerClass.warning;
   } else if (anyAreBelowThreshold(lineItems)) {
     status = InvoiceBannerStatus.invoiceBelowThreshold;
-    bannerClass = InvoiceBannerClass.notice;
+    bannerClass = InvoiceBannerClass.payable;
   } else if (anyAreNotNoCharge(lineItems)) {
     status = InvoiceBannerStatus.invoiceWithinThreshold;
     bannerClass = InvoiceBannerClass.payable;
