+++ src/frontend/utils/index.ts
@@ -1,5 +1,5 @@
 export * from "./months";
 export * from "./status-label";
-export * from "./percentage-discrepancy-special-case";
+export * from "./percentage-discrepancy-special-cases";
 export * from "./links";
 export * from "./title";
